fix structure code
