+++ src/App.js
@@ -1,13 +1,10 @@
-import React, { useEffect } from "react"
+import React from "react"
 import "./styles/App.css"
 import "./styles/loading.css"
 import { BrowserRouter} from "react-router-dom";
 import Routes from "./routes";
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 import { GRP_COLOR} from "./constant/css_constant"
-import { user_verify } from "./features/auth";
-import { useDispatch } from "react-redux";
-import SignIn from "./views/pages/auth/Login";
 
 function App() {
