+++ src/constant/css_constant.js
@@ -54,4 +54,4 @@ export const BOX_SHADOW = {
     CODE001: "0px 4px 4px rgba(0, 0, 0, 0.3)",
 }
 
-export const DRAWER_WITH = 240;
+export const DRAWER_WITH = 350;
