+++ src/views/auth/RegisterConfirm.js
@@ -2,8 +2,8 @@ import { Box, Button, Typography } from "@mui/material";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
-import { BORDER_RADIUS, BOX_SHADOW, GRP_COLOR } from "../../../constant/css_constant";
-import { registerConfirm } from "../../../features/auth";
+import { BORDER_RADIUS, BOX_SHADOW, GRP_COLOR } from "../../constant/css_constant";
+import { registerConfirm } from "../../features/auth";
