+++ src/views/auth/RegisterEmailSendSuccess.js
