+++ src/views/auth/ResetPwdEmailConfirm_backup.js
