+++ src/views/auth/ResetPwdEmailSendSuccess.js
