+++ src/views/chat/areaChat/message/PairingPersonMsg.js
