+++ src/views/chat/areaChat/message/SystemMessage.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ChatLayer from './message-description/ChatLayer';
 
-export default function AdminMsg() {
+export default function SystemMessage() {
   return (
     <>
       <div>AdminMsg</div>
