+++ src/views/chat/areaChat/sendMessageBox/TextChatLayer.js
