+++ src/views/chat/leftBar/LeftSideBar.js
@@ -1,0 +1,17 @@
+import { Box } from "@mui/system";
+import React from "react";
+import DrawerSideBar from "../../common/drawer/DrawerSideBar";
+
+
+export default function LeftSideBar() {
+  return (
+    <Box>
+      {/* <TopSideBar/>
+      <SearchChatBoxLayer />
+      <GroupChatSideBar />
+      <ChatSideBar /> */}
+
+      <DrawerSideBar />
+    </Box>
+  );
+}
