+++ src/views/chat/leftBar/components/ChatInfoLayer.js
