+++ src/views/chat/topBar/TopBar.js
@@ -1,15 +1,19 @@
 import { AppBar } from "@mui/material";
 import React from "react";
-import AppBarComponent from "../../../../common/drawer/component/AppBarComponent";
-import { DRAWER_WITH } from "../../../../../constant/css_constant";
+import { DRAWER_WITH } from "../../../constant/css_constant";
+import AppBarComponent from "../../common/drawer/component/AppBarComponent"
+import PartnerSettingModal from "../popup/PartnerSettingModal"
+
 export default function TopBar() {
+
   return (
     <AppBar
       position="fixed"
       elevation={0}
       sx={{
-        width: { sm: `calc(100% - ${DRAWER_WITH}px)` },
+        // width: { sm: `calc(100% - ${DRAWER_WITH}px)` },
         ml: { sm: `${DRAWER_WITH}px` },
+       borderBottom: "1px solid #e0e0e0",
       }}
     >
       <AppBarComponent />
