+++ src/views/common/base/emoji/EmojiLayer.js
