--- src/views/pages/chat/components/areaChat/message/MyMsg.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import ChatLayer from './message-description/ChatLayer';
-
-export default function MyMsg() {
-  return (
-    <>
-      <div>MyMsg</div>
-      <ChatLayer />
-    </>
-  );
-}
